+++ CHANGELOG.md
@@ -14,7 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 - Fix file parsing without a new line at the end.
-
+- Function arguments are case insensitive.
 
 ## [0.4.5] - 2023-05-29